Fix ItemsetSupport to correctly report support.
ItemsetSupport was using a hashmap with an integer key mapping to the
support. The integer key was the hashcode of the items array. This was
incorrect, as multiple arrays can hash to the same hash code, so when
collisions occurred in the hash table, the old support value would be
overwritten, causing multiple itemsets to have the same supports, and one
would be incorrect.

So change to using an ArrayList as the key.